Refactor select_server functions to static methods
